requirements.txt
pytest
pytest-asyncio
httpx
pydantic
pydantic-settings
allure-pytest
faker
jsonschema
pytest-xdist
eval_type_backport
pyotp
